notebooks/lesson_17/telegrambot_part1[20.01.25]/webservice/requirements.txt
uvicorn==0.34.0
fastapi==0.115.6
python-dotenv==1.0.1
langchain-community==0.3.13
langchain-openai==0.2.14
faiss-cpu==1.9.0.post1

# site
django == 5.1.4
django-cors-headers == 4.6.0

# telegram
python-telegram-bot==21.10
